Now the backend will send the progress event and the graphical part looks ok now too...

+++ desktop/src/main/nl/tudelft/ti2806/riverrush/screen/PlayingGameScreen.java
@@ -247,7 +247,7 @@ public class PlayingGameScreen implements Screen {
 
     }
 
-    public void updateProgress(int teamID, int progress) {
+    public void updateProgress(int teamID, double progress) {
         this.midScreen.updateProgress(teamID,progress);
     }
 }
